+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Edits a game.
 - Don't verify emails in dev.
 - Adds a delete game button.
+- Adds person icon to the join button.
 
 ## 0.18.0 [2023-11-04]